clean up nomad files
