+++ core/src/main/scala/scheduler/NomadHttp.scala
@@ -18,15 +18,15 @@ package nelson
 package scheduler
 
 import nelson.Datacenter.{Deployment, StackName}
-import nelson.Json._
 import nelson.Manifest._
 import nelson.blueprint.{ContextRenderer, EnvValue, Render}
 import nelson.blueprint.DefaultBlueprints.magnetar
-import nelson.docker.Docker
 import nelson.docker.Docker.Image
+
 import cats.~>
 import cats.effect.IO
 import cats.implicits._
+
 import java.util.concurrent.ScheduledExecutorService
 
 import journal.Logger
@@ -41,7 +41,7 @@ object NomadHttp {
       import Render.keys._
       import EnvValue._
 
-      override def inject(context: Magnetar.Context): Map[String, EnvValue] = {
+      override def render(context: Magnetar.Context): Map[String, EnvValue] = {
         import context._
         Map(
           (vaultPolicies, ListValue(
@@ -163,230 +163,12 @@ final class NomadHttp(
   def parse(dc: Datacenter, spec: String): IO[Json] = {
     log.debug(s"normalizing the following spec: $spec")
     val req = addCreds(dc, Request[IO](Method.POST, nomad.endpoint / "v1" / "jobs" / "parse"))
-    client.expect[Json](req.withBody(parseReqJson(spec)))
+    client.expect[Json](req.withBody(makeParseRequest(spec)))
   }
 
   def call(name: String, dc: Datacenter, json: Json): IO[String] = {
     log.debug(s"sending nomad the following payload: ${json.nospaces}")
     val req = addCreds(dc, Request[IO](Method.POST, nomad.endpoint / "v1" / "job" / name))
     client.expect[String](req.withBody(json))
-  }
-}
-
-object NomadJson {
-  import argonaut._, Argonaut._
-  import argonaut.EncodeJsonCats._
-  import Infrastructure.Nomad
-  import scala.concurrent.duration._
-  import NomadHttp._
-
-  sealed abstract class NetworkMode(val asString: String)
-  final case object BridgeMode extends NetworkMode("bridge")
-  final case object HostMode extends NetworkMode("host")
-
-  // We need to pass in the id because Nomad uses it as a key in the response :(
-  def deploymentSummaryDecoder(id: String): DecodeJson[DeploymentSummary] =
-    DecodeJson[DeploymentSummary](c => {
-      val inner = c --\ "Summary" --\ id
-      for {
-        running   <- (inner --\ "Running").as[Option[Int]]
-        failed    <- (inner --\ "Failed").as[Option[Int]]
-        queued    <- (inner --\ "Queued").as[Option[Int]]
-        completed <- (inner --\ "Complete").as[Option[Int]]
-      } yield DeploymentSummary(running, queued, completed, failed)
-    })
-
-  // reference: https://www.nomadproject.io/docs/http/jobs.html
-  // Note: implicit from argonaut Decoder[List[A]] => Decoder[A] is already defined
-  implicit def runningUnitDecoder: DecodeJson[RunningUnit] =
-    DecodeJson[RunningUnit](c => {
-      for {
-        nm <- (c --\ "Name").as[Option[String]]
-        st <- (c --\ "Status").as[Option[String]]
-        p  <- (c --\ "ParentID").as[Option[String]]
-      } yield RunningUnit(nm, st, p)
-    })
-
-  def parseReqJson(spec: String): argonaut.Json = {
-    argonaut.Json(
-      "JobHCL"       := spec,
-      "Canonicalize" := true
-    )
-  }
-
-  def dockerConfigJson(
-    nomad: Nomad,
-    container: Docker.Image,
-    ports: Option[Ports],
-    nm: NetworkMode,
-    name: String,
-    ns: NamespaceName): argonaut.Json = {
-
-    val maybePorts = ports.map(_.nel.map(p => argonaut.Json(p.ref := p.port)))
-
-    ("port_map" :=? maybePorts) ->?: argonaut.Json(
-    "image" := s"https://${container.toString}", // https:// required to work with private repo
-    "network_mode" := nm.asString,
-    "auth" := List(argonaut.Json(
-       "username" := nomad.dockerRepoUser,
-       "password" := nomad.dockerRepoPassword,
-       "server_address" := nomad.dockerRepoServerAddress,
-       "SSL" := true
-     ))
-    )
-  }
-
-  // cpu in MHZ, mem in MB
-  def resourcesJson(cpu: Int, mem: Int, ports: Option[Ports]): argonaut.Json = {
-    val maybePorts = ports.map(_.nel.map(p => argonaut.Json(
-      "Label" := p.ref,
-      "Value" := 0 // for dynamic ports this is required but is then ignored, garbage
-    )))
-    argonaut.Json(
-      "CPU"      := cpu,
-      "MemoryMB" := mem,
-      "IOPS"     := 0,
-      "Networks" := List(
-        ("DynamicPorts" :=? maybePorts) ->?: argonaut.Json(
-        "mbits" := 1 // https://github.com/hashicorp/nomad/issues/1282
-      ))
-    )
-  }
-
-  def logJson(maxFiles: Int, maxFileSize: Int): argonaut.Json = {
-    argonaut.Json(
-      "MaxFiles"      := maxFiles,
-      "MaxFileSizeMB" := maxFileSize
-    )
-  }
-
-  def vaultJson(ns: NamespaceName, name: String): argonaut.Json = {
-    // Names need to be namespaced, because a secret in dev != a secret in prod
-    val policyName = getPolicyName(ns, name)
-    ("Policies" := List(policyName)) ->:
-    ("Env" := true) ->:
-    ("ChangeMode" := "restart") ->:
-    ("ChangeSignal" := "") ->:
-    jEmptyObject
-  }
-
-  def healthCheckJson(check: HealthCheck): argonaut.Json = {
-    // to make nomad happy
-    // type can be http or tcp
-    // protocol can be http or empty
-    val (typ,protocol): (String, Option[String]) =
-      if (check.protocol == "http" || check.protocol == "https")
-        ("http", Some(check.protocol))
-      else (check.protocol, None)
-
-    val skipVerify =
-      if (check.protocol == "https") true else false
-
-    argonaut.Json(
-      "Name" := check.name,
-      "PortLabel" := check.portRef,
-      "Path":= check.path.getOrElse(""),
-      "Protocol":= protocol,
-      "Interval":= check.interval.toNanos,
-      "Timeout":= check.timeout.toNanos,
-      "TLSSkipVerify" := skipVerify,
-      "Type":= typ,
-      "Args":= argonaut.Json.jNull,
-      "Id" := "",
-      "Command":= ""
-    )
-  }
-
-  // Kind crude, we need better health checks
-  def servicesJson(name: String, port: Port, tags: Set[String], checks: List[HealthCheck]): argonaut.Json = {
-    val checksJson =
-      if (checks.isEmpty) // default tcp check if no health check is defined
-        List(healthCheckJson(HealthCheck(s"tcp ${port.ref} ${name}", port.ref, "tcp", None, 10.seconds, 4.seconds)))
-      else
-        checks.map(healthCheckJson)
-
-    argonaut.Json(
-      "Name" := name,
-      "PortLabel" := port.ref,
-      "Tags" := tags,
-      "Checks" := checksJson
-    )
-  }
-
-  def ephemeralDiskJson(sticky: Boolean, migrate: Boolean, size: Int): argonaut.Json = {
-    argonaut.Json(
-      "Sticky":= sticky,
-      "Migrate":= migrate,
-      "SizeMB":= size
-    )
-  }
-
-  def envJson(bindings: List[EnvironmentVariable]): argonaut.Json =
-    bindings.foldLeft(jEmptyObject)((j,a) => (a.name := a.value) ->: j)
-
-  def periodicJson(expression: String): argonaut.Json = {
-    ("Spec" := expression) ->:
-    ("Enabled" := true) ->:
-    ("SpecType" := "cron") ->:
-    ("ProhibitOverlap" := true) ->:
-    jEmptyObject
-  }
-
-  def restartJson(retries: Int): argonaut.Json = {
-    argonaut.Json(
-      "Interval":= 5.minutes.toNanos,
-      "Attempts":= retries,
-      "Delay" := 15.seconds.toNanos,
-      "Mode" := "delay"
-    )
-  }
-
-  def leaderTaskJson(name: String, unitName: UnitName, i: Image, env: Environment, nm: NetworkMode, ports: Option[Ports], nomad: Nomad, ns: NamespaceName, plan: PlanRef, tags: Set[String]): argonaut.Json = {
-    // Nomad does not support resource requests + limits so we just use limits here
-    val cpu = (nomad.mhzPerCPU * env.cpu.limitOrElse(0.5)).toInt
-    val mem = env.memory.limitOrElse(512.0).toInt
-    val services = ports.map(_.nel.map(p => servicesJson(
-      name,
-      p,
-      Set(ns.root.asString, s"port--${p.ref}", s"plan--$plan").union(tags),
-      env.healthChecks.filter(_.portRef == p.ref)
-    )))
-    ("Services" :=? services) ->?:
-    argonaut.Json(
-      "Vault"     := vaultJson(ns, name),
-      "Name"      := name,        // Maybe use Static names (primary, sidecar)
-      "Driver"    := "docker",
-      "leader"    := true,          // When "true", other tasks (necessarily "non leaders") will be gracefully terminated, when the leader task completes
-      "Config"    := dockerConfigJson(nomad, i, ports, nm, name, ns),
-      "Env"       := envJson(env.bindings),
-      "Resources" := resourcesJson(cpu, mem, ports),
-      "LogConfig" := logJson(10,10)
-    )
-  }
-
-  def job(name: String, unitName: UnitName, plan: Plan, i: Image, dc: Datacenter, schedule: Option[Schedule], ports: Option[Ports], ns: NamespaceName, nomad: Nomad, tags: Set[String]): Json = {
-    val env = plan.environment
-    val periodic = schedule.flatMap(_.toCron().map(periodicJson))
-    val scheduler = if (schedule.isDefined) "batch" else "service"
-    argonaut.Json(
-      "Job" :=
-        ("Periodic"   :=? periodic) ->?: argonaut.Json(
-        "Region"      := dc.name,
-        "Datacenters" := List(dc.name), // regions each have a single eponymous datacenter
-        "ID"          := name,
-        "Name"        := name,
-        "Type"        := scheduler,
-        "Priority"    := 50,
-        "AllAtOnce"   := false,
-        "TaskGroups"  := List(
-          argonaut.Json(
-            "Name"          := name,
-            "Count"         := env.desiredInstances.getOrElse(1),
-            "RestartPolicy" := env.retries.map(restartJson).getOrElse(restartJson(3)), // if no retry specified, only try 3 times rather than 15.
-            "Tasks"         := List(leaderTaskJson(name,unitName,i,env,BridgeMode,ports,nomad,ns,plan.name, tags))
-          )
-        )
-      )
-    )
   }
 }
