+++ core/src/main/scala/scheduler/package.scala
@@ -20,7 +20,7 @@ package object scheduler {
   // a hydrated blueprint (i.e. passes manifest validation and exists
   // in the database) so we simply take the supplied plan and extract
   // the `Blueprint`, and `Template` in turn.
-  def mkTemplate(env: Manifest.Environment, fallback: => IO[Template]): IO[Template] =
+  def mkTemplate(env: Manifest.Environment, fallback: IO[Template]): IO[Template] =
     env.blueprint match {
       case Some(Left((ref, rev))) => IO.raiseError(UnhydratedBlueprint(ref, rev))
       case Some(Right(bp))        => IO.pure(bp.template)
